backend/requirements.txt
arrow==1.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
gTTS==2.5.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
pygame==2.6.1
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.3
six==1.17.0
speedtest-cli==2.1.3
types-python-dateutil==2.9.0.20250516
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
